decouple tools pkg from filepathfilter
